feat: Move ITask and IMeeting to interfaces folder
