+++ src/models/meeting.ts
@@ -1,14 +1,5 @@
-import mongoose, { Document, Schema } from 'mongoose';
-
-export interface IMeeting extends Document {
-  userId: string;
-  title: string;
-  date: Date;
-  participants: string[];
-  transcript: string;
-  summary: string;
-  actionItems: string[];
-}
+import mongoose, { Schema } from 'mongoose';
+import { IMeeting } from '../interfaces/meeting.interfaces';
 
 const meetingSchema = new Schema<IMeeting>({
   userId: String,
@@ -17,7 +8,7 @@ const meetingSchema = new Schema<IMeeting>({
   participants: [String],
   transcript: String,
   summary: String,
-  actionItems: [String]
+  actionItems: [String],
 });
 
 export const Meeting = mongoose.model<IMeeting>('Meeting', meetingSchema);
